rename reducers and actions files
