+++ src/applications/personalization/dashboard/reducers/index.js
@@ -2,7 +2,7 @@ import claimsAppeals from '../../../claims-status/reducers';
 import prescriptions from '../../../rx/reducers/prescriptions';
 import recipients from '../../../messaging/reducers/recipients';
 import folders from '../../../messaging/reducers/folders';
-import preferences from '../../preferences/reducers';
+import preferences from '../../preferences/reducers/reducers';
 import appointments from '../../appointments/reducers';
 import authorization from 'applications/disability-benefits/686/reducers/authorization';
